Disable Secret File System link expiry validation
Since Drupal does not support adding cache metadata from
StreamWrappers it can currently occur that we cache a link in a render
cache longer than the link is valid which would cause viewers to be
provided with broken links. To avoid this we currently just assume that
a link with a valid hash is secret enough and thus always valid for some
time in the future. There is no current path to adding cache context in
StreamWrappers so there's no single issue to point to, but this is a
combination of various cache API and file system improvements that
people are discussing. Once fixed, uncomment the next line and remove
the one after.

Once https://www.drupal.org/project/drupal/issues/3358113 is fixed the
SecretStream can be amended and this commit can be reverted to re-enable
expiry checking.