Provide an update hook to migrate old group data
This provides an upgrade hook that uses the Drupal Entity system to
inspect the tables and create a single migration transaction. The use of
entity system ensures we catch all fields and take any database config
(such as table prefixes) into account. Additionally it should ensure
that we also catch data for custom plugins or fields.

Everything is wrapped in a migration to ensure we either succeed or are
put back in a known stable state, and not left hanging midway through a
migration.